fix bug in questions

+++ src/pages/QuestionDetails/QuestionDetails.jsx
@@ -2,13 +2,12 @@ import { useState, useEffect } from "react"
 import { useParams, Link } from "react-router-dom"
 import styles from './QuestionDetails.module.css'
 import * as questionService from '../../services/questionService'
-
 import Loading from "../Loading/Loading"
-//import AuthorInfo from "../../components/AuthorInfo/AuthorInfo"
+
 
 const QuestionDetails = (props) => {
   const { id } = useParams()
-  const [question, setQuestion] = useState([])
+  const [question, setQuestion] = useState(null)
 
   useEffect(() => {
     const fetchQuestion = async () => {
@@ -18,22 +17,27 @@ const QuestionDetails = (props) => {
     fetchQuestion()
   }, [id])
 
-// if (!question) return <Loading />
+if (!question) return <Loading />
 
   return (
     <main className={styles.container}>
       <article>
+          {console.log('author here', question.author._id)}
         <header>
           <h1>DETAILS</h1>
           <h3>{question.question}</h3>
           <span>
-            {question.author._id === props.user.profile? 
-              <>
+          {/* <AuthorInfo content={question} /> */}
+            {question.author._id === props.user.profile && 
+          <>
                 <Link to={`/questions/${id}/edit`} state={question}>Edit</Link>
+
                 <button onClick={() => props.handleDeleteQuestion(id)}>Delete</button>
-              </> : null
+
+              </> 
+
+  
             }
-            {console.log("AUTHOR", question.author)}
           </span>
         </header>
         <p>{question.text}</p>
